Reorganize files out of pir/ into src/ and src/lib
There is one small issue created by this commit, in that the plumage
binary is created in the src/ directory instead of the root of the
plumage checkout. pbc_to_exe does not take an output file option,
so the file will have to be copied/moved to the root of the repo
in a platform independent way.