[t] Plumage gets to be the first kid on the block to use the Tapir test harness
Tapir is just three PIR files, t/harness.pir and two libraries in
ext/Tapir/. A PBC or fakecutable for Tapir could be generated, but it
complicates the build process for no good reason. The make target for
'test' now invokes Tapir and there is now also a 'testv' target which is
verbose, i.e. prints out all TAP output.

In theory, we could add the Tapir git repo as a git submodule, but the current
setup is the simplest possible setup that could work.